pyproject.toml
[tool.poetry]
name = "mlt"
version = "0.1.0"
description = "ml tool for modelling"
authors = ["Rudabeh Meskarian"]
packages = [
    { include = "mlt"},
]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.21.0"
pandas = "^2.2.0"
scikit-learn = "^1.0.0"
#torch = "^1.9.0"
matplotlib = "^3.4.0"
pyyaml = "^6.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
shap = "^0.46.0"
# pyarrow = "^17.0.0"
seaborn = "^0.13.2"
xgboost = "^2.1.4"
# pmdarima = "2.0.4"
ray = "^2.47.1"


[tool.poetry.extras]
xgboost = ["xgboost"]

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.6b0"
flake8 = "^3.9.0"
ipykernel = "^6.16.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black] 
line-length = 120 
extend-exclude = "notebooks/*"

[tool.flake8]
max-line-length = 120

[tool.isort] 
line_length = 120 
multi_line_output = 5

[tool.poetry2conda]
name = "mlt"
default-channel = "pip"